fixed rounding errors

+++ LuckParser/Models/Statistics.cs
@@ -136,10 +136,10 @@ namespace LuckParser.Models
         {
             public int HitCount { get; }
             public int TotalHitCount { get; }
-            public int DamageGain { get; }
+            public double DamageGain { get; }
             public int TotalDamage { get; }
 
-            public DamageModifierData(int hitCount, int totalHitCount, int damageGain, int totalDamage)
+            public DamageModifierData(int hitCount, int totalHitCount, double damageGain, int totalDamage)
             {
                 HitCount = hitCount;
                 TotalHitCount = totalHitCount;
